+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/Plutus/TxInfo.hs
@@ -20,9 +20,6 @@
 {-# OPTIONS_GHC -Wno-orphans #-}
 
 module Cardano.Ledger.Alonzo.Plutus.TxInfo (
-  EraPlutusTxInfo (..),
-  EraPlutusContext (..),
-  mkPlutusLanguageContext,
   ContextError (..),
   TxOutSource (..),
   transDataHash,
@@ -55,7 +52,8 @@ import Cardano.Crypto.Hash.Class (hashToBytes)
 import Cardano.Ledger.Address (RewardAcnt (..))
 import Cardano.Ledger.Allegra.Scripts (ValidityInterval (..))
 import Cardano.Ledger.Alonzo.Era (AlonzoEra)
-import Cardano.Ledger.Alonzo.Scripts (AlonzoEraScript, PlutusScript (..))
+import Cardano.Ledger.Alonzo.Plutus.Context
+import Cardano.Ledger.Alonzo.Scripts (PlutusScript (..))
 import Cardano.Ledger.Alonzo.Tx (ScriptPurpose (..))
 import Cardano.Ledger.Alonzo.TxBody (
   AlonzoEraTxBody (..),
@@ -82,8 +80,7 @@ import Cardano.Ledger.Mary.Value (
   MultiAsset (..),
   PolicyID (..),
  )
-import Cardano.Ledger.Plutus.Data (Data (..))
-import Cardano.Ledger.Plutus.Language (Language (..), PlutusLanguage)
+import Cardano.Ledger.Plutus.Language (Language (..))
 import Cardano.Ledger.Plutus.TxInfo
 import Cardano.Ledger.PoolParams (PoolParams (..))
 import Cardano.Ledger.SafeHash (hashAnnotated)
@@ -99,7 +96,6 @@ import Control.Arrow (left)
 import Control.DeepSeq (NFData)
 import Data.ByteString.Short as SBS (fromShort)
 import Data.Foldable (Foldable (..))
-import Data.Kind (Type)
 import qualified Data.Map.Strict as Map
 import Data.Maybe (mapMaybe)
 import qualified Data.Set as Set
@@ -107,69 +103,7 @@ import Data.Text (Text)
 import GHC.Generics (Generic)
 import Lens.Micro ((^.))
 import NoThunks.Class (NoThunks)
-import qualified PlutusLedgerApi.V1 as P (ToData, toData)
 import qualified PlutusLedgerApi.V1 as PV1
-
-class (PlutusLanguage l, AlonzoEraScript era) => EraPlutusTxInfo (l :: Language) era where
-  toPlutusTxCert :: proxy l -> TxCert era -> Either (ContextError era) (PlutusTxCert l)
-
-  toPlutusScriptPurpose ::
-    proxy l ->
-    ScriptPurpose era ->
-    Either (ContextError era) (PlutusScriptPurpose l)
-
-  toPlutusTxInfo ::
-    proxy l ->
-    PParams era ->
-    EpochInfo (Either Text) ->
-    SystemStart ->
-    UTxO era ->
-    Tx era ->
-    Either (ContextError era) (PlutusTxInfo l)
-
-  toPlutusScriptContext ::
-    proxy l ->
-    PlutusTxInfo l ->
-    ScriptPurpose era ->
-    Either (ContextError era) (PlutusScriptContext l)
-
-class
-  ( AlonzoEraScript era
-  , Eq (ContextError era)
-  , Show (ContextError era)
-  , NFData (ContextError era)
-  , NoThunks (ContextError era)
-  , EncCBOR (ContextError era)
-  , DecCBOR (ContextError era)
-  ) =>
-  EraPlutusContext era
-  where
-  data ContextError era :: Type
-
-  mkPlutusScriptContext ::
-    PlutusScript era ->
-    ScriptPurpose era ->
-    PParams era ->
-    EpochInfo (Either Text) ->
-    SystemStart ->
-    UTxO era ->
-    Tx era ->
-    Either (ContextError era) (Data era)
-
-mkPlutusLanguageContext ::
-  (EraPlutusTxInfo l era, P.ToData (PlutusScriptContext l)) =>
-  proxy l ->
-  ScriptPurpose era ->
-  PParams era ->
-  EpochInfo (Either Text) ->
-  SystemStart ->
-  UTxO era ->
-  Tx era ->
-  Either (ContextError era) (Data era)
-mkPlutusLanguageContext proxy scriptPurpose pp epochInfo sysStart utxo tx = do
-  txInfo <- toPlutusTxInfo proxy pp epochInfo sysStart utxo tx
-  scriptContext <- toPlutusScriptContext proxy txInfo scriptPurpose
-  pure $ Data $ P.toData scriptContext
 
 instance Crypto c => EraPlutusTxInfo 'PlutusV1 (AlonzoEra c) where
   toPlutusTxCert _ = pure . transTxCert
