+++ eras/alonzo/test-suite/src/Test/Cardano/Ledger/Alonzo/Translation/TranslatableGen.hs
@@ -16,7 +16,7 @@ module Test.Cardano.Ledger.Alonzo.Translation.TranslatableGen (
 ) where
 
 import Cardano.Ledger.Alonzo (Alonzo)
-import Cardano.Ledger.Alonzo.Plutus.TxInfo
+import Cardano.Ledger.Alonzo.Plutus.Context
 import Cardano.Ledger.Alonzo.Scripts (AlonzoEraScript (eraMaxLanguage))
 import Cardano.Ledger.Core as Core
 import Cardano.Ledger.Plutus.Language (Language (..), SLanguage (..))
