+++ eras/alonzo/test-suite/test/Test/Cardano/Ledger/Alonzo/TxInfo.hs
@@ -9,7 +9,8 @@ import Cardano.Ledger.Address (Addr (..), BootstrapAddress (..))
 import Cardano.Ledger.Allegra.Scripts (ValidityInterval (..))
 import Cardano.Ledger.Alonzo (Alonzo)
 import Cardano.Ledger.Alonzo.Core hiding (TranslationError)
-import Cardano.Ledger.Alonzo.Plutus.TxInfo (toPlutusTxInfo, transVITime)
+import Cardano.Ledger.Alonzo.Plutus.Context (toPlutusTxInfo)
+import Cardano.Ledger.Alonzo.Plutus.TxInfo (transVITime)
 import Cardano.Ledger.Alonzo.Tx (AlonzoTx (..), IsValid (..))
 import Cardano.Ledger.Alonzo.TxBody (AlonzoTxBody (..), AlonzoTxOut (..))
 import Cardano.Ledger.BaseTypes (Network (..), StrictMaybe (..), natVersion)
