+++ eras/babbage/impl/src/Cardano/Ledger/Babbage/Rules/Utxos.hs
@@ -18,11 +18,11 @@ module Cardano.Ledger.Babbage.Rules.Utxos (
   babbageEvalScriptsTxInvalid,
 ) where
 
+import Cardano.Ledger.Alonzo.Plutus.Context (EraPlutusContext)
 import Cardano.Ledger.Alonzo.Plutus.Evaluate (
   collectPlutusScriptsWithContext,
   evalPlutusScripts,
  )
-import Cardano.Ledger.Alonzo.Plutus.TxInfo (EraPlutusContext)
 import Cardano.Ledger.Alonzo.Rules (
   AlonzoUtxosEvent (..),
   AlonzoUtxosPredFailure (..),
