+++ libs/cardano-ledger-test/src/Test/Cardano/Ledger/Examples/AlonzoCollectInputs.hs
@@ -14,6 +14,7 @@
 module Test.Cardano.Ledger.Examples.AlonzoCollectInputs (tests) where
 
 import Cardano.Ledger.Alonzo (Alonzo)
+import Cardano.Ledger.Alonzo.Plutus.Context (mkPlutusScriptContext)
 import Cardano.Ledger.Alonzo.Plutus.Evaluate (CollectError (..), collectPlutusScriptsWithContext)
 import Cardano.Ledger.Alonzo.Plutus.TxInfo
 import Cardano.Ledger.Alonzo.Scripts (AlonzoEraScript (..), AlonzoScript (..), ExUnits (..))
